Use configuration_aliases to define alternative providers
Previous versions of Terraform required an empty "proxy configuration
block" to declare multiple configuration names for a provider within a
module. Terraform 0.15 adds a new field to the required_providers block
for providers to indicate aliased configuration names. Using the
configuration_aliases argument resolves warnings emitted when this
module is used on newer versions of Terraform.